new code to train models
this is one script to train any gpt-style model with hyperparameters that align to the Raspberry Pi's hardware constrains. It will normally train DistilGPT2, but there are other two models that can be trained with the same code. If you want to train other model, then uncomment or comment out accordingly as is presented in the code